GH-1080 - first step, define a factory interface, initially for retrieving deserializers and a KafkaConsumerFactory mostly taken from DefaultKafkaConsumerFactory, which makes use of the factory.
At this stage (once tests have been written) any clients could use the new factory if they provide their own implementation of the KafkaSerializerDeserializerFactory.java. The next step
is to refactor DefaultKafkaConsumerFactory so it delegates to the new type.

Overall aim of fix for GH-1080 is to ensure KafkaConsumers and KafkaProducers created by factories can have their own instances of Serializers and Deserializers so that their lifecycles can be aligned.