GH-1080 tidy up some of the warnings.
However, using the annotation approach has complications - if the serializers/deserializers are to be limited to specific consumerfactories, then those consumerFactories have to be declared as beans. The problem then is that if you do not declare a listenerContainerFactory named "kafkaListenerContainerFactory" then one is created in KafkaAnnotationDrivenConfiguration and that requires one and only one bean of type consumerFactory. In my opinion, the configuration should probably be based on a missing type rather than name, but changing that could have knock on impacts. As such, the rules around using annotations for serializers and deserializers are more complex regarding what you can and can't do, and this is covering quite a niche use case, better not to introduce complexity. Will now fall back on the original plan of allowing a supplier of Serializer/Deserializer.